test(resource): split client test files
